requirements.txt
requests~=2.25
sentry-sdk~=0.19
praw~=7.1
python-telegram-bot~=13.0
youtube-dl==2021.2.10
